types: Split v19 blockchain module up
Split the module up as we do for others. Done in preparation for adding
support for `getblockfilter`.

Note we use the same public re-export of errors with a wildcard. Its a
know problem I still need to fix crate wide.
